move HathiTrust::SIP::YAML.load to HathiTrust::SIP::SIP.load_yaml
Although it did actually work as written, it was very confusing to
debug.
